[matching] Consider the match valid if the 2 first matches correspond to the "same" feature
In [Mishkin 2015] "MODS: Fast and Robust Method for Tow-View Matching"
they introduce the FGINN strategy "First Geometrically Inconsistent
Nearest Neighbor". Here, for simplicity and performances, we don't
search for the first geometrically inconsistent but assume that the
match is strong enough if the 2 nearest neighbors are on the same image
region.
It could be 2 really close features or the same feature with multiple
descriptors (different orientations of SIFT for instance).