[software] cmdline: avoid usage of bool_switch
bool_switch only allows to enable the argument, so it was unusable for
all options with "true" as the default value.
As it only allows to enable, the argument must be named accordingly, so
it sometimes ends up with negation in the name.
And finally it makes things a bit more complex to generate the command
line from scripts.