[delunaycut] create mesh from delaunayGC: new option to filter triangles containing helper points without creating holes
Some vertices have not been created by depth maps but
have been created for the tetrahedralization like
camera centers, helper points, etc.
These points have no visibility information (nrc == 0).
We want to remove these fake points, but we don't want to create holes.
So if the vertex is alone in the middle of valid points, we want to keep
it.