Update cameraSensors.db (#505)
* Update cameraSensors.db

Adding support for the LG-V30.  This is the model I own and tested.  Passes Camera init test.  I am not able to complete render as I do not have an nvidia card that is cuda capable to try.  Here is sony spec for the camera: https://www.sony-semicon.co.jp/products_en/IS/sensor1/img/products/ProductBrief_IMX351_20171109.pdf

Also this is for the main camera only not the wide camera