[sfmDataIO] Alembic: use XForms instead of untyped objects for root nodes
There is no standard way to handle Alembic nodes without type information (OObject). Using OXForm guarantees that the generated Alembic file can be properly loaded in other applications, and with the correct hierarchy.
This does not introduce compatibility issues: the AlembicImporter does not require root nodes to be untyped objects to parse an Alembic sfmData file.