[mesh] subdivide a simplified mesh with uv coords update
Modify functions to subdivide a mesh to update uv coords, and use it to texture a simplified mesh.
WARNING: for now, all meshes with uv coords are subdivided & mesh subdivision assumes that the mesh already has uv coords. The texturing of a subdivided mesh only works for meshes with one texture atlas