[ci] Ground truth evaluation: Use image filename as UID
The UID generation has been updated, so if there is no unique ID in the
image metadata, the full path of the image will be used.
This breaks the ways we generate identical UID for GT evaluation.
So with this PR, the GT comparaison uses the number extracted from the
filename to create the UID.