[fuseCut] DelaunayGraphCut: Use directly isLineInTriangle (CRITIC)
Use directly isLineInTriangle in getFacetFromVertexOnTheRayToTheCam instead of rayCellIntersection because we come directly from a vertex and only the opposite facet of an adjacent cell can be intersected.

Stop the collision tests as soon as an intersection with a facet is found.
The configuration with convex geometry of the adjacent 
 cells(tetrahedrons) to a vertex means that only two intersections can be found with a ray crossing this vertex.
When the intersection verifying the distance condition is found, it can be returned without iterating on the other cells.