Add dcp color profile loading and applying capabilities when raw images must be read.
In cameraInit node, using the maker and model metadata, if a color profile database is provided, the complete filepath of the appropriate dcp file is added in the image metadata. Then, every time the image has to be read, the color profile can be applied. Application options are available. As a consequence if a node wants to get the original raw images, this is still possible by deactivating the color profile application.
A new parameter in cameraInit node allows to select the error level (warning or error) when a database is provided but no adapted color profile is found for a raw input image.