Improve test reproducibility
Many tests call Vec*::Random() and similar functions from Eigen, which
make the tests not reproducible. This is problematic, because the tests
become unstable and even if some of them sometimes fail indicating a
problem, it's hard to debug and fix the failure as the exact conditions
of the failing test are not known. Initializing random number generators
to a static seed fixes the issue.