Add option to compile Eigen with alignment enabled
Eigen is most optimal when using SIMD instructions which require
overaligned buffers. In C++14 this is a problem because there's no
built-in support for correctly aligned allocations, so this Eigen
feature is currently disabled in AliceVision. C++17 solves Eigen
alignment issues completely because there now exist additional overloads
of operator new() which support correct allocation alignment. We can't
upgrade to C++17 yet, but certain compilers added support for this
feature earlier, such as -faligned-new in GCC 7.1. Therefore it makes
sense to add a configuration option to allow enabling alignment support
in Eigen in cases when the compiler is good enough.