[system] Implement ProgressDisplay wrapper for boost::progress_display
This wrapper is intended to abstract the exact way progress is
displayed. This will be useful later as a way to inject the exact
progress display implementation depending on how aliceVision is being
used. For example, if aliceVision is used as a library then it makes
more sense to give progress information back to the caller instead of
formatting and sending it to console.