[CMake] Add option to build dependencies in parallel
Currently all dependencies are built using single CPU core for each
dependency. Presumably the previous idea was that parallelization comes
from AliceVision itself being built in parallel so even if each
dependency uses just a single CPU core, the dependencies themselves are
built in parallel and the impact is not that large.

However, ability to run compilation within each dependency in parallel
is still useful, especially when the build machine has a large number of
cores, or distributed compilation is used.