Fix broken openmp atomic accesses
OMP_ATOMIC_UPDATE and OMP_ATOMIC_WRITE are currently completely broken.

Currently OMP_ATOMIC_UPDATE and OMP_ATOMIC_WRITE are defined as follows:

> #define OMP_ATOMIC_UPDATE _Pragma("omp atomic update")
> #define OMP_ATOMIC_WRITE  _Pragma("omp atomic write")

These macros are used as follows:

> #pragma OMP_ATOMIC_UPDATE
> #pragma OMP_ATOMIC_WRITE

Which results in the following code after preprocessor expansion:

> #pragma _Pragma("omp atomic update")
> #pragma _Pragma("omp atomic write")

This is invalid pragma which is ignored by the compiler. The correct
usage of _Pragma is without #pragma as follows:

> _Pragma("omp atomic update")
> _Pragma("omp atomic write")

Fixing this issue reveals another problem that "omp atomic" is very
restrictive in what expressions it accepts. In many cases the
expressions include access to std::vector and so on, which "omp atomic"
does not support.

This problem is addressed by using boost::atomic_ref which allows to
apply atomic operations on any variable. Because "omp atomic" is hard to
use, it is replaced with boost::atomic_ref in all cases.