[mvsData] Have a separate function to convert EImageColorSpace for OIIO
Currently EImageColorSpace_enumToString is used for two purposes: as an
enum formatter in cases such as command line argument display and
parsing; and as a mapping to OpenImageIO identifiers used for
conversion. It makes sense to have two separate (even if identical)
functions for this purpose as it is then clear what use case is and what
guarantees expect from their behavior.