[mvsData] Merge imageIO::EImageColorSpace to image::EImageColorSpace
Behavior should remain the same except that the code in many cases
supports only a subset of EImageColorSpace depending on whether it was
originally written to target imageIO::EImageColorSpace or
image::EImageColorSpace.

The only potential change is that the default colorspace names are now
lowercase in cases where imageIO::EImageColorSpace was used before.
While AliceVision ignores the case when parsing colorspace, this could
affect external tools.