[voctree] Run kmeans trials in parallel
Currently each kmeans trial is run sequentially which is later
parallelized across all cores when summing distances. This has higher
overhead compared to performing all trials in parallel and then
parallelizing for summing because less threads are started up. numTrials
is currently 5, so the number of thread startups is up to 5 times less
than before.