[pipeline] Fix and harmonize the descriptions of the command lines
Fix the executable's name if needed and harmonize the description if
there is any (description first, ended with a period and followed by
the executabl's name on a new line). Some descriptions are rephrased
to reduce the differences between executables. In case the description
or the name of the command line was erroneous, it is fixed.