[keyframe] Handle missing frames within an input video
If a frame is missing in a video sequence, instead of throwing an exception
straight away, try reading the next frame. If the next frame is valid, then
push dummy scores for the missing frame, and keep processing the input
video. Otherwise, do throw the exception and stop the process.

The dummy scores will be ignored in the final keyframe selection
(explicitly in the case of the motion accumulation computation, implicitly
when applying the weights during the sharpness selection).