[image] Add a function determining if the input extension is a video format
"isSupported()" relies exclusively on the content of OpenImageIO's
"extension_list", which contains all the formats supported as inputs,
including video formats, with no distinction.

A function "isVideoExtension" is added to determine whether the input
extension is part of the OIIO-supported movie formats. The list of
supported formats is hard-coded based on OIIO's documentation.