KeyframeSelection: Add debug option to skip sharpness score computation
Add a "skipSharpnessComputation" debug option that allows to compute the
scores with performing the sharpness score computations. All frames will
be assigned a fixed sharpness score of 1.0.

If the smart selection is applied with this option enabled, the selected
frames will be those located in the center of each subset (determined with
the motion scores) as the weights will be applied on constant scores.

This option is useful to determine the impact of the sharpness score
computation on the global processing time.