[dataio] Use 4-spaces indentation for all the module's files
This commit only modifies the syntax of the dataio module and does not
change its content in any way. The hpp and cpp files have been clang-
formatted to respect the 4-spaces indentation that can be found in other
modules (in the "keyframe" module, primarily), and some functions'
prototypes have been re-aligned accordingly.