[software] PanoramaCompositing: Update comment on the Gaussian filter
The Gaussian filter window used to be of size (1 + 2 * 2), with 2 the
radius of said Gaussian filter. With the change of the common scale from
the min scale to the max scale, it is updated to 1, but we might still
want to go back to the previous one later on.

The const variable gaussianFilterRadius is commented out since it is not
used anymore, but might be again in the future.