[software] PanoramaPostProcessing: Fall back to default preview size if it is set to 0
This commit adds a safeguard in case the user provides a panorama preview
width of 0. In that case, we should fall back to the default preview size
(1000). Otherwise, the whole process will go through with no preview being
generated, and an error will occur at the end when attempting to write the
non-existing image.