[keyframe] Fill the "selectedFrames" vector for the regular method
The "selectedFrames" vector is filled with 0s and 1s depending on whether
the frame with the ID corresponding to the position within the vector has
been selected as a keyframe or not.

It was filled with the smart method, but not with the regular one. As it
will be needed later on and is not costly, it should also be filled for
the regular selection method.