[keyframe] Write SfMData files as outputs if the input is an SfMData file
This commit adds a new method that writes two output SfMData files (one
containing the keyframes, the other containing all the frames that were not
selected) when the input file is an SfMData file.

Once all the selected keyframes have been written on disk, the input
SfMData file is loaded (outside of the feed) and its views and intrinsics
are extracted to be written in their corresponding output SfMData file.
No information written is these files is built from scratch.

Rigs are currently not supported and therefore ignored when writing the
output files.

If the input is a video or a sequence of images, no SfMData file is written
for now.