[software] CameraCalibration: Use latest factorized commmand line parsing
The parsing of the command line is now handled in another module and the
`aliceVision::CmdLine` object can be used for the command line. This
automatically adds the support for the memory- and core-related options.

Additionally, this commit updates some of the parameters' descriptions and
logs, as well as error handling.