[nodes] Publish: Add support to publish directories as well as files
Before this commit, the `Publish` node could only copy files to an output
directory. It can now copy a complete directory to an output directory.
This is useful for cases where the names or number of the files to publish
are unknown prior to the computation of the node to publish.