[lightingEstimation] Add safeguards for images that have no sphere
If there are some images in the input SfMData for which no sphere has
been detected, attempting to retrieve their sphere parameters in the tree
will cause a crash.

This commit checks that the node corresponding to an input image exists
before attempting to retrieve its parameters: if it does not, said node
is skipped and the image is not considered as an image to process during
the lighting calibration.

When writing the final JSON file that will contain the lighting models, the
list of images that do contain a sphere is provided as well as the SfMData.
This way, the file will only contain images containing spheres for which
the computation of a lighting model is valid.

Additionally, some variables in snake case have been renamed.