[sphereDetection] Exclude images with no sphere from the JSON file
When writing the final JSON file that sums up the detected spheres for each
input image, we do not include images for which no sphere has been
detected, to avoid filling the JSON with information that cannot be used.

Later on, during the lighting calibration part, images from the input
SfMData that cannot be found in the sphere detection's JSON will be
ignored, which will prevent issues in further stages.