[photometricStereo] Do not make up mask name when it is not provided
If no mask path is provided as an input, an empty string should be passed
to the `loadMask` function rather than making up a path.

When entering the `loadMask` function, there is now a distinction between
a mask that wasn't provided and a mask that doesn't exist. In both cases,
this leads to using a mask in which all the pixels are set to 1, but the
corresponding log will differ (in the first case, it is expected, while
it is not in the second, thus triggering a warning).