[photometricStereo] Prevent crash when no sphere was detected for an image
If no sphere has been detected for an input image, its light has not been
calibrated, and building its light intensities' matrix will result in an
empty matrix.

During the photometric stereo, the light intensities are directly accessed
without any checks, as it is assumed that at this point, the light must
have been calibrated. If the matrix is empty, accessing it will cause a
crash.

This commit prevents this by checking beforehand that the size of the
matrix of light intensities corresponds to the size of input images. If
this is not the case, the image is skipped with a warning message.

By the end of the process, if all the images have been skipped, an error
is thrown, since it means that either no input was meant for the
photometric stereo or something wrong occurred during the previous stages.