[photometricStereo] Issue warning if no image shared the same pose ID
If the all the processed images were included in lists of size 1, it means
that none of them shared the same pose ID. To share the same pose ID,
images must be grouped together in a folder starting with "ps_" when they
are provided to the `CameraInit` node.

If there was no group of images, this might mean that images were not
placed in correct folders, or that all the processed images were
independent. A warning is issued to let the user know that the results
might not be the expected ones because of the location of the input images.