[keyframe] Don't write keyframes on disk if the output extension is `none`
Add the support for a `none` value for the output extension value. If this
value is provided, the selected keyframes will not be written on disk at
the end of the process. Only the SfMData files (one containing the
keyframes, one containing the frames that were not selected) will be
generated.

In the case of video inputs, which already only produce the SfMData file
containing the keyframes, setting the output extension to `none` will lead
to no output at all, since the keyframes' SfMData file needs them to be
written on disk.

The README is also updated with that information.