[software] Do not fail if range start is larger than the number of views
For cases where a node declares a size that is bigger than it actually is,
the start of ranges might exceed the number of views in the SfMData.

Instead of causing an EXIT_FAILURE, we detect that the size of the range
is incorrect to proceed with computations, and return EXIT_SUCCESS.

This allows to use nodes such as the `KeyframeSelection`, for which the
number of outputs can rarely be known before the end of the computations:
if the number of chunks for `PrepareDenseScene` and `FeatureExtraction` is
too large for the actual size of `KeyframeSelection`, then we keep on
going through the pipeline instead of failing.