[ci] Update Github actions for the Windows build of the CI
Instead of downloading and rebuilding the vcpkg dependencies at every run,
this commit downloads a zip file containing all the prebuilt binairies.
The `VCPKG_ROOT` environment variable is set explicitly to ensure that
the vcpkg commands can be used without any issue.

The link to the zip file should be updated every time a dependency is
updated or added/removed, since a new archive should have been generated.

The CUDA toolkit has been re-enabled and its version should always be
aligned with the one used to generate the vcpkg archive (in this case,
11.6.0).

The build step itself has been modified as follows:
- the toolchain is explicitly provided as a compilation flag instead of
relying on the `useVcpkgToolchainFile` parameter
- the vcpkg triplet is explicitly provided
- the samples are not built anymore since they will not be tested anyway
- CUDA is enabled