[nodes] HDR Fusion: Detect new exposure groups when inputs' paths change
In addition to comparing the exposure levels of the sorted input images
to determine whether a new exposure group must be created, this commit
adds a detection based on the path of the image: if the directory
containing the current image is not the same as the one of the previous
image, then we assume the dataset is different, and there is no point
in comparing the current image's exposure level with the previous one's.

Instead, we directly create a new exposure group.

In cases where images from different datasets have very similar
exposure levels, this prevents having outliers identified as being part
of the current exposure group.