[nodes] HDR Fusion: Fix bracket detection
Prior to this commit, only the shutter speed was compared between two
images to determine whether they belonged to the same group. The fnumber
and ISO were assumed to be fixed within a group, which is not always true,
and differs from what is done on the AliceVision's side.

This commit aligns Meshroom's bracket detection with AliceVision's.