KeyframeSelection: Add new param to use masks for the smart selection
A list of paths to directories containing masks to apply during the
computation of the scores can now be provided as an input.

If a camera rig is being used and masks are provided, then the number of
directories containing masks must be the same as the number of media
paths.