[keyframe] Apply masks to the frames for the sharpness score computation
If masks have been provided and have passed all the checks (size of the
masks, number of masks, non-emptiness), apply them to the input frames
before the sharpness score is computed.

If a mask covers at least 50% of an evaluated window, it is ignored
and we go on to sliding the window.