[keyframe] Apply masks to images for the optical flow score computation
The optical flow is computed, and the mask is then applied on the
integral of the motion vectors. The summed motion values will then be
equal to 0 wherever the mask is applied.

When computing the norm of the cell, if we detect that at least 50% of the
cell is masked (by counting the number of 0s within the cell)