[utils] ColorCheckerCorrection: Add a default value when retrieving colors
In some cases, the colors written in the JSON file of the detected checkers
can be set as "-nan". When that occurs, the value is not considered as a
"double", which raises a fatal issue ("cannot converted type to double").

Here, if the requested color value cannot be retrieved, or if it is not
a double, it will be set to 0 instead of stopping the program.