[sfmDataIO] Convert string with boost::lexical_cast
`std::stoi` throws an `std::out_of_range` exception if the provided string
exceeds the maximum value for an integer (2147483647). When reading views,
it is quite common to have images that contain numbers larger than an
integer (images may be named with the date and time at which they were
taken, for example), thus causing an unhandled error.

Instead of converting the string to an integer, it should be converted
using boost::lexical_cast, which will throw a `boost::bad_lexical_cast`
exception if it cannot be converted.

In that case, the number is not extracted from the image's filename, and
the behaviour is the same as if there had not been any number in it.