[image] io: Rely on OIIO to determine if an image file exists or not
An image file's existence was checked prior to being passed to OIIO to
be loaded, and if the test failed, an error was thrown. This worked
fine for Linux platforms but in case a non-ASCII character was part of
the image's path on Windows, that test always failed because non-ASCII
characters are not correctly handled by default.

OIIO is however able to correctly read existing image files, independently
from the type of characters their path contains. If a file with or without
non-ASCII characters does not exist, the `ImageBuf` object will not be
correctly initialized, and we will be able to catch it and throw an error.
If it does exist, it will be correctly read in all cases.

There is thus no need for the test on the file's existence anymore.