[ci] Windows: Set `ALICEVISION_LIBPATH` to "bundle/bin"
`ALICEVISION_ROOT` is used to locate the installation folder pre-bundle.
Once the bundle has been performed, no clean-up step is done to "merge"
the bundle output with the installation folder, so `$ALICEVISON_ROOT/bin`
does not contain all the dependencies requested by the executables it
contains.

In order to load the DLLs needed by the Python unit tests, we set
`ALICEVISION_LIBPATH` to `$ALICEVISION_ROOT/bundle/bin`, which contains
all the needed dependencies.