PATCH: Use NEW behavior of CMake policy CMP0117
CMake used to pass /GR (RTTI) as a CXX flag to MSVC. This will be removed with CMP0117. AliceVision however depends on RTTI due to usage of dynamic_cast() and typeid. Although MSVC has /GR (RTTI) enabled by default, ensure it is explicitly passed as a compiler flag for clarity.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>